requirements/python2.txt
six
pytest
hathilda
requests
unicodecsv
beautifulsoup4
python-dateutil
